basic transition to cmake build + some reorganization of files
