--- debug/unittests/run.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-for folder in ut*
-do
-    cp run_single_unittest.sh ${folder}/
-    cd ${folder}
-    echo "-> running unittest ${folder}"
-    echo ""
-    ./run_single_unittest.sh
-    cd ..
-done
