+++ src/CMakeLists.txt
@@ -1,0 +1,2 @@
+file(GLOB SOURCES "*.cpp")
+add_library(mci STATIC ${SOURCES})
