--- src/mci
@@ -1,1 +1,0 @@
-./
