+++ test/README.md
@@ -1,7 +1,6 @@
 # LEGEND OF THE UNIT TESTS
 
 
-
 ## Unit Test 1
 
 `ut1/`: check that calling the integrator again without doing the findMRT2step and initialDecorrelation gives the same results.
