+++ test/ut1/main.cpp
@@ -1,5 +1,5 @@
-#include "MCIntegrator.hpp"
-#include "MCISamplingFunctionInterface.hpp"
+#include "mci/MCIntegrator.hpp"
+#include "mci/MCISamplingFunctionInterface.hpp"
 
 #include <iostream>
 #include <math.h>
@@ -59,21 +59,18 @@ int main(){
 
     // this integral will give a wrong answer! This is because the starting point is very bad and initialDecorrelation is skipped (as well as the MRT2step automatic setting)
     mci->setX(x);
-    mci->integrate(NMC, average, error, 0, 0);
+    mci->integrate(NMC, average, error, false, false);
     assert( abs(average[0]-CORRECT_RESULT) > 2.*error[0] );
 
     // this integral, instead, will provide the right answer
     mci->setX(x);
-    mci->integrate(NMC, average, error, 10, 1000);
+    mci->integrate(NMC, average, error);
     assert( abs(average[0]-CORRECT_RESULT) < 2.*error[0] );
 
     // now, doing an integral without finding again the MRT2step and doing the initialDecorrelation will also result in a correct result
-    mci->integrate(NMC, average, error, 0, 0);
+    mci->integrate(NMC, average, error, false, false);
     assert( abs(average[0]-CORRECT_RESULT) < 2.*error[0] );
 
-    // and using fixed blocking also gives the same result
-    mci->integrate(NMC, average, error, 0, 0, 15);
-    assert( abs(average[0]-CORRECT_RESULT) < 2.*error[0] );
 
 
     delete pdf;
